requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
python-dotenv
pandas==2.2.2
geopy==2.4.1
geopandas==1.0.1
numpy==2.0.0
scipy==1.13.1
azure-appconfiguration==1.6.0
azure-storage-blob==12.20.0
azure-storage-file-datalake==12.15.0